[김태현] 문제은행9

+++ TaehyunKim/Spring/demo/src/main/java/com/example/demo/entity/PB0903/RectangleCalculator.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class RectangleCalculator {
 
-    private float area;
+    private int area;
 
     public RectangleCalculator(int length, int height){
         area = length * height;
